skillora_project/requirements.txt
sqlalchemy
alembic
psycopg2-binary
python-dotenv
curl_cffi
apscheduler
cloudscraper
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
prometheus-fastapi-instrumentator
redis
pytest
pytest-cov
httpx
ruff
